refactor(input): wire up `useForm`
